backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-multipart==0.0.9
pandas==2.2.2
openpyxl==3.1.5
docxtpl==0.20.2